requirements.txt
sentence_transformers
transformers
numpy
websocket_client
loguru
openai